feat: refactor wallet logic and update connect wallet styles
